+++ CHANGELOG.md
@@ -5,6 +5,7 @@ adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Expose `notice.parsed_backtrace` (#454)
+- Use integrations' native error handlers over `Rails.error` where available, to avoid loss of context (#460)
 
 ## [5.0.2] - 2022-11-04
 ### Fixed